+开发下载功能中 ++调用系统内建下载器。 ++构建调用内置下载器 +++下载记录 room +瞎改逻辑

+++ Ti/app/src/main/java/com/example/kiylx/ti/downloadCore/DownloadServices.java
@@ -82,7 +82,7 @@ public class DownloadServices extends Service {
             return -1;
         }
         @Override
-        public long getRate(DownloadInfo info){
+        public float getRate(DownloadInfo info){
            return mDownloadManager.getPercentage(info);
 
         }
